feat: refactor label handling by renaming labeler.ts to labels.ts and updating related imports and tests
